pyproject.toml
[project]
name = "fringemanager"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel Weston"}
]
readme = "README.md"
requires-python = "=3.11.7"
dependencies = [
    "numpy (==2.3.0)",
    "pywin32 (>=311,<312)",
    "pillow (>=12.0.0,<13.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
